Stage 3: Page Requirements Limitation for not logged users

+++ Philately/src/main/java/com/philately/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.philately.controller;
 
+import com.philately.config.UserSession;
 import com.philately.model.entity.Stamp;
 import com.philately.service.StampService;
 import com.philately.service.UserService;
@@ -15,10 +16,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class UserController {
 
     private final UserService userService;
+    private final UserSession userSession;
     private final StampService stampService;
 
     @PostMapping("/add-to-wishlist/{stampUUID}")
     public String addToWishlist(@PathVariable String stampUUID) {
+        if (!this.userSession.isUserLogged()) {
+            return "index";
+        }
+
         Stamp stamp = this.stampService.getStampById(stampUUID);
         this.userService.addStampToWishList(stamp);
         return "redirect:/";
@@ -26,12 +32,20 @@ public class UserController {
 
     @PostMapping("/remove-from-wishlist/{stampUUID}")
     public String removeFromWishlist(@PathVariable String stampUUID) {
+        if (!this.userSession.isUserLogged()) {
+            return "index";
+        }
+
         this.userService.removeFromWishlist(stampUUID);
         return "redirect:/";
     }
 
     @PostMapping("/by-all")
     public String byAllStamps() {
+        if (!this.userSession.isUserLogged()) {
+            return "index";
+        }
+
         this.userService.byAllStamps();
         return "redirect:/";
     }
